tests/wpt/web-platform-tests/clipboard-apis/async-write-blobtext-read-blobtext-manual.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>
  Async Clipboard write ([text/plain Blob]) -> read ([text/plain Blob]) tests
</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async function readWriteTest(textInput) {
  promise_test(async t => {
    const blobInput = new Blob([textInput], {type: 'text/plain'});

    await navigator.clipboard.write([blobInput]);
    const blobsOutput = await navigator.clipboard.read();
    assert_equals(blobsOutput.length, 1);
    const blobOutput = blobsOutput[0];
    assert_equals(blobOutput.type, "text/plain");

    const textOutput = await (new Response(blobOutput)).text();
    assert_equals(textOutput, textInput);
  }, "Verify write and read clipboard given text: " + textInput);
}

readWriteTest("Clipboard write ([text/plain Blob]) -> read ([text/plain Blob]) test");
readWriteTest("non-Latin1 text encoding test データ");
</script>
<p>
  Note: This is a manual test because it writes/reads to the shared system
  clipboard and thus cannot be run async with other tests that might interact
  with the clipboard.
</p>
